bpf: add initial miniact implementation
This work adds an option for a native, class-less, bpf based qdisc,
which can be used as a lightweight drop-in option for ingress or
egress points. It supports single or multi-prog mode as well as
hardware offload for nfp in the first case. In the latter we always
run the attached programs to completion and similarly as in seccomp-bpf,
multi-prog bpf tracing and cgroup case enforce the stronger verdict
from all the programs.

Currently unspec/ok/drop verdicts are supported for the multi-prog
case; more can be enabled later on as well. The concept of miniact
is to have a contained, built-in model with BPF for the miniq and
it piggy-backs on its infrastructure. This approach also avoids the
use of retpolines that prevents from speculation and is easy to
integrate into orchestrators managing the devices. Support for nfp
and netdevsim has been implemented and tested as well as integration
and stress testing into Cilium.

Example output with nfp:

  [root@apoc]# tc qdisc show dev enp2s0
  qdisc mq 0: root
  qdisc fq_codel 0: parent :4 limit 10240p flows 1024 quantum 1514 target 5.0ms interval 100.0ms memory_limit 32Mb ecn
  qdisc fq_codel 0: parent :3 limit 10240p flows 1024 quantum 1514 target 5.0ms interval 100.0ms memory_limit 32Mb ecn
  qdisc fq_codel 0: parent :2 limit 10240p flows 1024 quantum 1514 target 5.0ms interval 100.0ms memory_limit 32Mb ecn
  qdisc fq_codel 0: parent :1 limit 10240p flows 1024 quantum 1514 target 5.0ms interval 100.0ms memory_limit 32Mb ecn
  qdisc miniact ffff: parent ffff:fff1 gen 2 ingress hw ids 7 egress sw ids 8 9 10

Example from a Cilium dev:

  [root@apoc]# tc qdisc show dev lxcbd830
  qdisc noqueue 0: root refcnt 2
  qdisc miniact ffff: parent ffff:fff1 gen 3 ingress sw ids 595

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>